+++ tests/testthat/test_sim_discrete_time.r
@@ -264,3 +264,21 @@ test_that("works with formulas", {
   sim <- sim_discrete_time(dag, n_sim=100, max_t=50)
   expect_equal(round(mean(sim$data$D), 3), -0.833)
 })
+
+test_that("helpful error message formula error", {
+
+  dag <- empty_dag() +
+    node("A", type="rbernoulli") +
+    node_td("B", type="gaussian", formula= ~ -1 + A*2 + C*3)
+
+  expect_error(sim_discrete_time(dag, n_sim=100, max_t=3),
+            paste0("An error occured when interpreting the formula of ",
+                   "node 'B'. The message was:\nError: Error: column(s) not ",
+                   "found: [A, C]\nThis error may occur when one of the terms ",
+                   "in a supplied ", "formula does not match any variables ",
+                   "in the generated data.\n Please check whether all terms ",
+                   "in your supplied formula occur in the data generated up ",
+                   "to this point.\n The variables currently available in ",
+                   "data are:\n(Intercept), ATRUE"), fixed=TRUE)
+
+})